Finishing.

+++ main.py
@@ -11,12 +11,12 @@ from paths import fix_path
 from pipeline import process_one_object
 from restructure_files import move_images_and_build_full_relations
 
-DESIRED_FILE_COUNT = 10  # How many item files (tar.gzip) to download
+DESIRED_FILE_COUNT = 9  # How many item files (tar.gzip) to download
 OBJECTS_PER_GROUP = 1  # How many objects from each group to process
 IMAGES_PER_OBJECT = 1  # How many images per object to process
 
 distortion_levels = [
-    {"blur": 0, "noise": 0, "exposure": 1.0},  # No distortion
+    {"blur": 0, "noise": 0, "exposure": 1.0},
     {"blur": 4, "noise": 0, "exposure": 1.0},
     {"blur": 10, "noise": 0, "exposure": 1.0},
     {"blur": 0, "noise": 60, "exposure": 1.0},
@@ -31,6 +31,7 @@ distortion_levels = [
 ROOT_DIR = Path(os.getcwd())
 BASE_DB_PATH = fix_path(ROOT_DIR / "datasets" / "omniobject3d")
 DRIVE_LINK_LIST_PATH = fix_path(ROOT_DIR / "curated_drive_ids.txt")
+# DRIVE_LINK_LIST_PATH = fix_path(ROOT_DIR / "drive_file_ids.txt")
 RAW_IMAGES_PATH = BASE_DB_PATH / "images_raw"
 IMAGES_PATH = BASE_DB_PATH / "images"
 POINTCLOUD_ROOT = BASE_DB_PATH / "ply_16384" / "extracted" / "16384"
@@ -49,7 +50,7 @@ print("CUDA available:", torch.cuda.is_available())
 print("GPU:", torch.cuda.get_device_name(0) if torch.cuda.is_available() else None)
 
 ############# DOWNLOADING DATABASE FILES #############
-SKIP_DOWNLOAD = False
+SKIP_DOWNLOAD = True
 
 if not SKIP_DOWNLOAD:
     # Read desired file IDs
@@ -69,31 +70,33 @@ if not SKIP_DOWNLOAD:
 
 ############# LOADING/TESTING #############
 
-taus = [0.1, 0.2, 0.5]
+taus = [0.05, 0.075, 0.1]
 
 # Load relations
 grouped_data = load_dataset_relations(json_path=FINAL_RELATION_FILE_PATH)
 print("Loaded:", len(grouped_data.keys()), "objects")
 
-# Identify all images that will actually be processed
-images_to_keep = set()
-for group_objs in grouped_data.values():
-    for obj_id, fields in list(group_objs.items())[:OBJECTS_PER_GROUP]:
-        img_files = sorted(list(Path(fields["images"]).glob("*.png")))[:IMAGES_PER_OBJECT]
-        for img_file in img_files:
-            images_to_keep.add(img_file.resolve())
+prune_files = False
+if prune_files:
+    # Identify all images that will actually be processed
+    images_to_keep = set()
+    for group_objs in grouped_data.values():
+        for obj_id, fields in list(group_objs.items())[:OBJECTS_PER_GROUP]:
+            img_files = sorted(list(Path(fields["images"]).glob("*.png")))[:IMAGES_PER_OBJECT]
+            for img_file in img_files:
+                images_to_keep.add(img_file.resolve())
 
-# Delete unneeded images in the main image folder
-for img_file in Path(IMAGES_PATH).rglob("*.png"):
-    if img_file.resolve() not in images_to_keep:
-        img_file.unlink()
+    # Delete unneeded images in the main image folder
+    for img_file in Path(IMAGES_PATH).rglob("*.png"):
+        if img_file.resolve() not in images_to_keep:
+            img_file.unlink()
 
-# Delete empty folders after purging unused images
-for dirpath, dirnames, filenames in os.walk(IMAGES_PATH, topdown=False):
-    if not dirnames and not filenames:  # empty folder
-        Path(dirpath).rmdir()
+    # Delete empty folders after purging unused images
+    for dirpath, dirnames, filenames in os.walk(IMAGES_PATH, topdown=False):
+        if not dirnames and not filenames:  # empty folder
+            Path(dirpath).rmdir()
 
-print(f"[INFO] Kept {len(images_to_keep)} images, deleted unused images from {IMAGES_PATH}")
+    print(f"[INFO] Kept {len(images_to_keep)} images, deleted unused images from {IMAGES_PATH}")
 
 results = {}
 for group_name, group_objs in grouped_data.items():
